+++ src/WorkflowEngine/Retry/index.tsx
@@ -1,10 +1,10 @@
 import { h, FunctionComponent } from 'preact'
 import classNames from 'classnames'
-import { trackComponent } from '../../../Tracker'
+import { trackComponent } from '../../Tracker'
 import { Button } from '@onfido/castor-react'
-import theme from 'components/Theme/style.scss'
-import ScreenLayout from '../../Theme/ScreenLayout'
-import PageTitle from '../../PageTitle'
+import theme from '../../components/Theme/style.scss'
+import ScreenLayout from '../../components/Theme/ScreenLayout'
+import PageTitle from '../../components/PageTitle'
 import style from './style.scss'
 import { WithTrackingProps } from '~types/hocs'
 import type { StepComponentBaseProps } from '~types/routers'
