+++ src/WorkflowEngine/Retry/style.scss
@@ -1,4 +1,4 @@
-@import '../../Theme/constants';
+@import '../../components/Theme/constants';
 
 .container {
   display: flex;
