+++ src/WorkflowEngine/index.tsx
@@ -1,5 +1,8 @@
 export * from './engine'
 export * from './poller'
+export * from './Retry'
+export * from './utils/WorkflowTypes'
+
 import { EngineInterface, EngineProps, Engine } from './engine'
 
 export const workflowEngine = ({
