+++ src/components/Router/createWorkflowStepsHook.tsx
@@ -4,8 +4,8 @@ import { useCallback, useEffect, useRef, useState } from 'preact/hooks'
 import { formatStep } from '../../index'
 import { FlowVariants, NarrowSdkOptions, UrlsConfig } from '~types/commons'
 import { StepsHook, CompleteStepValue } from '~types/routers'
-import { poller, PollFunc, Engine } from '../WorkflowEngine'
-import type { WorkflowResponse } from '../WorkflowEngine/utils/WorkflowTypes'
+import { poller, PollFunc, Engine } from '~WorkflowEngine'
+import type { WorkflowResponse } from '~WorkflowEngine/utils/WorkflowTypes'
 import useUserConsent from '~contexts/useUserConsent'
 import { noop } from '~utils/func'
 import useActiveVideo from '~contexts/useActiveVideo'
